requirements.txt
boto3 == 1.26.90
botocore == 1.29.90
xgboost == 1.7.5
scikit-learn == 1.2.2
pandas == 2.0.2
ipython == 8.12
retry == 0.9.2
pytest == 7.0.0
mock == 5.0.2
numpy == 1.24.3
python-dotenv == 1.0.0
mlflow == 2.5.0
httpx == 0.24.1
prefect == 2.11.0
typing-extensions == 4.6.0
pydantic == 1.10.0
fastapi == 0.93
pandas == 2.0.2

pytest-cov == 2.6.0
pytest-mock == 3.6.1
ipykernel == 6.23.1
gunicorn == 20.1.0
flask ==2.3.2

tqdm==4.65.0
requests==2.31.0
joblib==1.3.1
pyarrow==12.0.1
psycopg==3.1.9
psycopg_binary==3.1.9
evidently==0.3.3
pylint==2.14.4
pre-commit==3.3.3
black==23.7.0
